% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternal_mortality.R
\docType{data}
\name{maternal_mortality}
\alias{maternal_mortality}
\title{Example data}
\format{
dataframe with 186 obs. of  23 variables
}
\usage{
maternal_mortality
}
\description{
Maternal Mortality dataset from World Bank Data
}
\keyword{datasets}
