% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBprep.R
\name{WBprep}
\alias{WBprep}
\title{Preparing World Bank Data}
\usage{
WBprep(df, colname)
}
\arguments{
\item{colname}{string - name of morality count column}

\item{data}{World Bank dataset to be formatted and pivoted}
}
\value{
pivoted dataset
}
\description{
Preparing World Bank Data
}
\examples{
WBprep(maternal_mortality, "maternal_mortality")
}
